added jenkins.bouncycastle.prioritizeJceProvider system property
Added a new (optional) boolean system property which affects how BouncyCastle
is registered as a JCE provider.  When `true`, the provider is inserted at the
beginning of the list (in 2nd position), instead of being added at the end.

It could help in situations where BC does a better job than the JDK. In
particular, it should solve issues with SSH connections to some SunOS servers,
which end up with this exception (with JSCH from the publish-over-ssh plugin):

```
java.security.InvalidAlgorithmParameterException: DH key size must be multiple
of 64, and can only range from 512 to 4096 (inclusive). The specific key size
2047 is not supported
```

See https://stackoverflow.com/a/43121515, where this idea comes from.
In this case, simply having BouncyCastle registered but as the last provider is
not good enough to solve the issue.

Probably it could have been a solution for
[JENKINS-35231](https://issues.jenkins-ci.org/browse/JENKINS-35231) too.

The system property added here is expected to be set on Jenkins main process.
Registration on agents will follow the same preference, and not a property of
the agent process.