[JENKINS-66180] be tollerant of the presence of the BouncyCastle FIPS provider
The most common way of running a JVM based app in FIPS compliance mode
is to use the Bouncy Castle FIPS provider.

This provider uses the same package structure as the regular one, but
lacks support for some algorithms and some classes are different).
Because of this if we just try and register the BouncyCastle PRovider as
normal we will end up
1. subverting the FIPS compliance of the JVM
2. likely failing due to invalid class / incompatable classes causing
   plugin initialisation errors, causing dependent plugins to also not
   initialise.

Dependant plugins can often (but not always) be adapted to use either
the FIPS apis or the regular ones (mostly this in an excersize in not
using deprecated APIs!) but also ensuring that no alrorithm is asked for
that is not FIPS compliant.   In either case this would need changes in
those plugins, but for the plugins that need no changes they should
function regardless, so now we do not unconditionally register any
provider in either the controller or the agent.

The first thing we do is check for the prescence of the BouncyCastle
FIPS provider (using the BCFIPS provider string), if that is present
then this plugin effectively becomes a no-op plugin with the exception
of the PEM helper class.

if the FIPS provider is not present then the plugin dynamically updates
the classpath to inject the regualr bouncycastle libraries.

for plugins that depend on this plugin no changes should be needed as
long as everything is running in non FIPS mode.

For the plugin to also work in the presence of the FIPS provider they
need to
1. not use any non fips approved algorithms
2. use defaults (for keystore end factories) etc rather than hard coding
   vvalues
3. not use deprecated API
4. self check at runtime that things are still working as ecpected.