Handle BC-FIPS related exception and propagate exception causes
When testing with BC-FIPS PEMEncodeable was failing in some unexpected
ways.

in the first if we could not convert due to a PKCSException or InvalidKeySpecException
then the cause was logged and a new exception without details was thrown.
This means the underlying cause is lost when using any exception in a
FormValidation.

In the second case when using a key that had too short of salt an
org.bouncycastle.crypto.fips.FipsUnapprovedOperationError was thrown
which being an error would ripple up and cause an unexpected error.

We now catch this error and wrap it in an UnrecoverableKeyException so
it can be handled by the caller